feat(acoustic): expose per-band/preset material ABI and add chord_functional_analysis + mel_ex APIs
C-ABI acoustic surface (sonare_c_acoustic.h/cpp, ABI version 2):
- SonareRirSynthConfig gains `absorption_bands / absorption_band_count /
  material_preset` fields; wall material selection now follows a three-tier
  precedence: named SONARE_MATERIAL_PRESET_* > per-band array > scalar
  absorption.
- SonareRoomMorphConfig gains the same material tail plus `mixing_time_ms`,
  `crossfade_ms`, and `late_model` (Sabine/Eyring) fields that were previously
  only reachable from C++.
- New SONARE_MATERIAL_PRESET_{CONCRETE,WOOD,CURTAIN,CARPET,GLASS} constants
  (mirrors MaterialPreset enum).
- Static layout asserts changed from sizeof-total to offsetof-prefix so the
  optional pointer tail does not break the check.

C-ABI feature surface (sonare_c_features.h/cpp, sonare_c.cpp):
- `sonare_chord_functional_analysis`: detect chords and label each with a
  Roman-numeral relative to a supplied key (key_root, key_mode); returns a
  heap-owned SonareStringArray.
- `sonare_mel_spectrogram_ex` / `sonare_mfcc_ex`: forward-transform variants
  with explicit fmin/fmax/htk so a caller can round-trip with the inverse
  (sonare_mel_to_stft / sonare_mel_to_audio) using the same Mel parameters.
- `sonare_free_string_array` exposed in the header.
- Output structs in sonare_analyze / sonare_analyze_impulse_response /
  sonare_detect_acoustic / sonare_audio_detect_onsets are now zero-initialized
  before the try block, so a rejected input leaves a fully-consistent result.

Feature layer (src/feature/):
- Minor correctness fixes feeding the above: chroma/CQT/VQT tuning alignment,
  Mel-spectrogram fmin/fmax threading, rhythm/segment/spectral defaults.

WASM bindings (src/wasm/bindings.cpp, bindings/wasm/):
- Wire all new C-ABI functions; expose SonareStringArray result type.
- Update TypeScript types (sonare.js.d.ts, index.ts) and add coverage tests.

Node bindings (bindings/node/):
- sonare_wrap_analysis.cpp: add functional_analysis and mel_spectrogram_ex/
  mfcc_ex wrappers; update index.ts + header + tests.

Python bindings (bindings/python/):
- _ffi.py: declare new FFI signatures; _analysis.py + _features.py: add
  chord_functional_analysis, mel_spectrogram_ex, mfcc_ex wrappers.
- engine.py / analyzer.py: expose the new APIs; update .pyi stubs.
- New and extended tests in test_analyzer.py, test_engine.py,
  test_new_bindings.py.

Parity tooling (tools/parity/allowlist.toml):
- Register mel_spectrogram_ex, mfcc_ex, and chord_functional_analysis as
  intentional binding-shape variants (ABI options-bag vs. native keyword args).

+++ src/feature/cqt.cpp
@@ -318,8 +318,8 @@ CqtResult cqt(const Audio& audio, const CqtConfig& config, CqtProgressCallback p
 
   using VectorXcf = Eigen::Matrix<std::complex<float>, Eigen::Dynamic, 1>;
 
-  // Pre-allocate output
-  std::vector<std::complex<float>> output(n_bins * n_frames);
+  // Pre-allocate output (promote to size_t before multiplying to avoid int overflow)
+  std::vector<std::complex<float>> output(static_cast<size_t>(n_bins) * n_frames);
 
   // Create FFT processor
   FFT fft(fft_length);
