+++ src/feature/vqt.cpp
@@ -290,8 +290,8 @@ VqtResult vqt(const Audio& audio, const VqtConfig& config, VqtProgressCallback p
 
   using VectorXcf = Eigen::Matrix<std::complex<float>, Eigen::Dynamic, 1>;
 
-  // Allocate output
-  std::vector<std::complex<float>> output(n_bins * n_frames);
+  // Allocate output (promote to size_t before multiplying to avoid int overflow)
+  std::vector<std::complex<float>> output(static_cast<size_t>(n_bins) * n_frames);
 
   // Create FFT processor
   FFT fft(fft_length);
